refactor(dashboard): decompose server hotspots for issue 384
